+++ core/application/MainWindowUIAdapter.h
@@ -25,7 +25,7 @@ public:
     void setDividedMeshVisibility(int meshIndex, bool visible) override;
     void setDividedMeshOpacity(int meshIndex, double opacity) override;
     void setDividedMeshFileName(int meshIndex, const QString& fileName) override;
-    std::vector<double> getStressThresholds() const override;
+    std::vector<int> getStressThresholds() const override;
     std::vector<StressDensityMapping> getStressDensityMappings() const override;
     QString getCurrentMode() const override;
     void setStressRange(double minStress, double maxStress) override;
